requirements.txt
appdirs==1.4.4
asgiref==3.3.4
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
cryptography==3.4.7
cycler==0.10.0
decorator==5.0.9
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.3
django-bootstrap4==3.0.1
django-colorfield==0.4.1
django-countries==7.2.1
django-crispy-forms==1.11.2
django-decouple==2.1
django-extensions==3.1.3
django-filter==2.4.0
django-guardian==2.4.0
django-multiselectfield==0.1.12
django-phonenumber-field==5.1.0
django-rest-auth==0.9.5
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
djangorestframework-serializer-extensions==2.0.1
Faker==8.7.0
filelock==3.0.12
hashids==1.3.1
httpie==2.4.0
idna==2.10
jwt==1.2.0
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.2
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.0.25
mysqlclient==2.0.3
numpy==1.20.3
pandas==1.2.4
phonenumbers==8.12.24
Pillow==8.2.0
plotly==4.14.3
protobuf==3.18.1
pycparser==2.20
Pygments==2.9.0
PyMySQL==1.0.2
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
razorpay==1.2.0
requests==2.25.1
requests-toolbelt==0.9.1
retrying==1.3.3
scipy==1.6.3
seaborn==0.11.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
text-unidecode==1.3
Unipath==1.1
urllib3==1.26.5
validators==0.18.2
virtualenv==20.4.7
